Combine Company Creation with Initial Location form
