+++ app/views/companies/new.html.erb
@@ -1,3 +1,4 @@
+<h2>Company Information</h2>
 <%= form_for(@company) do |f| %>
   <div class="form-group">
     <%= f.text_field :name %>
@@ -40,7 +41,7 @@
     </div>
   </div>
 
-  <!-- Fields for LOCATION -->
+  <h3>Company Primary Location</h3>
 
   <%= fields_for(@location) do |u| %>
     <div class="form-group">
@@ -78,12 +79,12 @@
       </div>
     </div>
 
-    <!-- <div class="form-group">
-      <%#select_tag :state, options_for_select(@location.us_states) %>
+    <div class="form-group">
+      <%= select_tag :state, options_for_select(@location.state_options) %>
       <div class="col-md-4">
-        <%# u.label :state %>
+        <%= u.label :state %>
       </div>
-    </div> -->
+    </div>
 
     <div class="form-group">
       <%= u.text_field :zip_code %>
